# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Book Management System API
  description: A comprehensive REST API for managing books, authors, publishers, and orders
  version: 1.0.0
  contact:
    name: API Support
    email: support@bookmanagement.com
  license:
    name: ISC
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.bookmanagement.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        role:
          type: string
          enum: [user, admin]
          example: user
        avatar:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        provider:
          type: string
          nullable: true
          example: google
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: FAIL
        message:
          type: string
          example: Error message
        statusCode:
          type: integer
          example: 400
...
    RegisterResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...
  responses:
...
    BadRequestError:
      description: Bad request - Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password. Users with admin emails will be assigned admin role.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...